+++ backend/users/urls.py
@@ -1,7 +1,7 @@
-from django.urls import re_path
+from django.urls import path, re_path
 
 from rest_framework.authtoken.views import obtain_auth_token
-from users.views import UserCreateView, UserView, UserEventView, UserCompanyView, UserAdminView, UserContactView
+from users.views import UserOrderView, UserCreateView, UserView, UserEventView, UserCompanyView, UserAdminView, UserContactView
 
 urlpatterns = [
     re_path(r"^$", UserView.as_view(), name="user"),
@@ -11,4 +11,5 @@ urlpatterns = [
     re_path(r"^events/", UserEventView.as_view(), name="user_event"),
     re_path(r"^companies/", UserCompanyView.as_view(), name="user_company"),
     re_path(r"^contacts/",  UserContactView.as_view(), name="user_contacts"),
+    path("<int:id>/orders/", UserOrderView.as_view(), name="user_orders")
 ]
